requirements.txt
Click==7.0
Flask==1.0.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
paho-mqtt==1.4.0
redis==3.0.1
Werkzeug==0.14.1
colorama==0.4.1
MarkupSafe==1.1.1
Pillow==5.4.1
qrcode==6.1
six==1.12.0
